# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "me-builder API",
    "description": "me-builder Web UIが利用するHTTP API契約",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "liffIdToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "LIFF ID token"
      }
    }
  },
  "paths": {
    "/api/line/liff/session": {
      "post": {
        "operationId": "verifyLiffSession",
        "tags": ["LIFF"],
        "summary": "LIFF IDトークンを検証してAccountを解決する",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idToken": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": ["idToken"],
                "$schema": "http://json-schema.org/draft-07/schema#"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "表示可能なLINEプロフィール",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "displayName": {
                      "type": "string",
                      "minLength": 1
                    },
                    "pictureUrl": {
                      "type": "string",
                      "format": "uri"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "401": {
            "description": "LIFF IDトークンを検証できない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Unauthorized"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "404": {
            "description": "対応するAccountが存在しない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Account not found"
                    },
                    "reason": {
                      "const": "friendship_required"
                    }
                  },
                  "required": ["error", "reason"]
                }
              }
            }
          },
          "500": {
            "description": "未処理のサーバーエラー",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Internal Server Error"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "503": {
            "description": "D1 bindingが設定されていない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Service Unavailable"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/statistics": {
      "get": {
        "operationId": "getAdminStatistics",
        "tags": ["Admin"],
        "summary": "GeminiとLINEの当月利用統計を取得する",
        "security": [
          {
            "liffIdToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "管理者向け利用統計",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "period": {
                      "type": "object",
                      "properties": {
                        "start": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "end": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": ["start", "end"]
                    },
                    "fetchedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "gemini": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "status": {
                              "const": "available"
                            },
                            "estimatedCostUsd": {
                              "type": "number"
                            },
                            "requestCount": {
                              "type": "number"
                            },
                            "inputTokens": {
                              "type": "number"
                            },
                            "outputTokens": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "status",
                            "estimatedCostUsd",
                            "requestCount",
                            "inputTokens",
                            "outputTokens"
                          ]
                        },
                        {
                          "type": "object",
                          "properties": {
                            "status": {
                              "const": "unavailable"
                            },
                            "reason": {
                              "enum": ["not-configured", "upstream-error"],
                              "type": "string"
                            }
                          },
                          "required": ["status", "reason"]
                        }
                      ]
                    },
                    "line": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "status": {
                              "const": "available"
                            },
                            "billableMessages": {
                              "type": "number"
                            },
                            "monthlyLimit": {
                              "anyOf": [
                                {
                                  "type": "number"
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            },
                            "replyMessages": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "status",
                            "billableMessages",
                            "monthlyLimit",
                            "replyMessages"
                          ]
                        },
                        {
                          "type": "object",
                          "properties": {
                            "status": {
                              "const": "unavailable"
                            },
                            "reason": {
                              "enum": ["not-configured", "upstream-error"],
                              "type": "string"
                            }
                          },
                          "required": ["status", "reason"]
                        }
                      ]
                    }
                  },
                  "required": ["period", "fetchedAt", "gemini", "line"]
                }
              }
            }
          },
          "401": {
            "description": "LIFF IDトークンを検証できない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Unauthorized"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "403": {
            "description": "管理者権限がない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Forbidden"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "404": {
            "description": "対応するAccountが存在しない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Account not found"
                    },
                    "reason": {
                      "const": "friendship_required"
                    }
                  },
                  "required": ["error", "reason"]
                }
              }
            }
          },
          "500": {
            "description": "未処理のサーバーエラー",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Internal Server Error"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "503": {
            "description": "D1 bindingが設定されていない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Service Unavailable"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnoses": {
      "get": {
        "operationId": "listDiagnoses",
        "tags": ["Diagnosis"],
        "summary": "回答進捗を含む診断一覧を取得する",
        "security": [
          {
            "liffIdToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "診断一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "diagnoses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "minLength": 1
                          },
                          "title": {
                            "type": "string",
                            "minLength": 1
                          },
                          "description": {
                            "type": "string",
                            "minLength": 1
                          },
                          "opensAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "closesAt": {
                            "anyOf": [
                              {
                                "type": "string",
                                "format": "date-time"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "displayOrder": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "availability": {
                            "enum": ["open", "closed"],
                            "type": "string"
                          },
                          "responseStatus": {
                            "enum": ["unanswered", "in-progress", "answered"],
                            "type": "string"
                          },
                          "answeredCount": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "questionCount": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 9007199254740991
                          },
                          "lastAnsweredAt": {
                            "anyOf": [
                              {
                                "type": "string",
                                "format": "date-time"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "description",
                          "opensAt",
                          "closesAt",
                          "displayOrder",
                          "availability",
                          "responseStatus",
                          "answeredCount",
                          "questionCount",
                          "lastAnsweredAt"
                        ]
                      }
                    }
                  },
                  "required": ["diagnoses"]
                }
              }
            }
          },
          "401": {
            "description": "LIFF IDトークンを検証できない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Unauthorized"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "404": {
            "description": "対応するAccountが存在しない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Account not found"
                    },
                    "reason": {
                      "const": "friendship_required"
                    }
                  },
                  "required": ["error", "reason"]
                }
              }
            }
          },
          "500": {
            "description": "未処理のサーバーエラー",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Internal Server Error"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "503": {
            "description": "D1 bindingが設定されていない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Service Unavailable"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnoses/{diagnosisId}": {
      "get": {
        "operationId": "getDiagnosisDetail",
        "tags": ["Diagnosis"],
        "summary": "新規回答用の診断詳細を取得する",
        "security": [
          {
            "liffIdToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Question VersionとChoiceを含む診断詳細",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 1
                    },
                    "title": {
                      "type": "string",
                      "minLength": 1
                    },
                    "description": {
                      "type": "string",
                      "minLength": 1
                    },
                    "opensAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "closesAt": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "questions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "diagnosisQuestionId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "questionId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "questionVersion": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 9007199254740991
                          },
                          "text": {
                            "type": "string",
                            "minLength": 1
                          },
                          "hint": {
                            "anyOf": [
                              {
                                "type": "string",
                                "minLength": 1
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "choices": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "choiceId": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "label": {
                                  "type": "string",
                                  "minLength": 1
                                }
                              },
                              "required": ["choiceId", "label"]
                            },
                            "minItems": 2,
                            "maxItems": 2
                          }
                        },
                        "required": [
                          "diagnosisQuestionId",
                          "questionId",
                          "questionVersion",
                          "text",
                          "hint",
                          "choices"
                        ]
                      },
                      "minItems": 1
                    }
                  },
                  "required": ["id", "title", "description", "opensAt", "closesAt", "questions"]
                }
              }
            }
          },
          "401": {
            "description": "LIFF IDトークンを検証できない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Unauthorized"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "404": {
            "description": "対応するAccountがない、またはDiagnosisが公開されていない",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "const": "Account not found"
                        },
                        "reason": {
                          "const": "friendship_required"
                        }
                      },
                      "required": ["error", "reason"]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "const": "Diagnosis not found"
                        },
                        "reason": {
                          "const": "diagnosis_not_found"
                        }
                      },
                      "required": ["error", "reason"]
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Diagnosisの受付が終了している",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Diagnosis closed"
                    },
                    "reason": {
                      "const": "diagnosis_closed"
                    }
                  },
                  "required": ["error", "reason"]
                }
              }
            }
          },
          "500": {
            "description": "未処理のサーバーエラー",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Internal Server Error"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "503": {
            "description": "D1 bindingが設定されていない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Service Unavailable"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "diagnosisId",
            "required": true
          }
        ]
      }
    },
...
    "/api/dev/diagnosis-data": {
      "delete": {
        "operationId": "resetDevelopmentDiagnosisData",
        "tags": ["Development"],
        "summary": "開発環境で本人の診断回答データを全削除する",
        "security": [
          {
            "liffIdToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "削除した本人の回答関連データ件数",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deletedResponseCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "deletedAnswerCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "deletedDeferredQuestionCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "deletedSourceRecordCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "deletedResponseCount",
                    "deletedAnswerCount",
                    "deletedDeferredQuestionCount",
                    "deletedSourceRecordCount"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "LIFF IDトークンを検証できない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Unauthorized"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "404": {
            "description": "開発環境ではない、または対応するAccountがない",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "const": "Account not found"
                        },
                        "reason": {
                          "const": "friendship_required"
                        }
                      },
                      "required": ["error", "reason"]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "const": "Not Found"
                        }
                      },
                      "required": ["error"]
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "未処理のサーバーエラー",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Internal Server Error"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "503": {
            "description": "D1 bindingが設定されていない",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "const": "Service Unavailable"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    }
  }
}